класс. регистр uint16_t -> uint32_t

+++ qengine/utils/types.h
@@ -32,7 +32,7 @@
 namespace qengine {
 inline namespace type {
 
-using CReg = uint16_t;
+using CReg = uint32_t;
 
 template <typename T>
 using RVec = std::vector<T>;
